fix: wrapped graphql data mutations in transaction

+++ backend/molgenis-emx2-graphql/src/main/java/org/molgenis/emx2/graphql/GraphqlTableFieldFactory.java
@@ -977,38 +977,46 @@ public class GraphqlTableFieldFactory {
   private DataFetcher fetcher(Schema schema, MutationType mutationType) {
     return dataFetchingEnvironment -> {
       StringBuilder result = new StringBuilder();
-      boolean any = false;
-      for (TableMetadata tableMetadata : schema.getMetadata().getTables()) {
-        List<Map<String, Object>> rowsAslistOfMaps =
-            dataFetchingEnvironment.getArgument(tableMetadata.getIdentifier());
-        if (rowsAslistOfMaps != null) {
-          String tableName = tableMetadata.getTableName();
-          Table table = tableMetadata.getTable();
-          int count;
-          List<Row> rows = TypeUtils.convertToRows(table.getMetadata(), rowsAslistOfMaps);
-          switch (mutationType) {
-            case UPDATE:
-              count = table.update(rows);
-              result.append("updated " + count + " records to " + tableName + "\n");
-              break;
-            case INSERT:
-              count = table.insert(rows);
-              result.append("inserted " + count + " records to " + tableName + "\n");
-              break;
-            case SAVE:
-              count = table.save(rows);
-              result.append("upserted " + count + " records to " + tableName + "\n");
-              break;
-            case DELETE:
-              boolean strict = dataFetchingEnvironment.getArgumentOrDefault("strict", false);
-              count = table.delete(rows, strict);
-              result.append("delete " + count + " records from " + tableName + "\n");
-              break;
-          }
-          any = true;
-        }
+      if (schema.getMetadata().getTables().stream()
+          .noneMatch(t -> dataFetchingEnvironment.getArgument(t.getIdentifier()) != null)) {
+        throw new MolgenisException("None or invalid tables provided");
       }
-      if (!any) throw new MolgenisException("Error with save: no data provided");
+      schema
+          .getDatabase()
+          .tx(
+              db -> {
+                Schema txSchema = db.getSchema(schema.getName());
+                for (TableMetadata tableMetadata : txSchema.getMetadata().getTables()) {
+                  List<Map<String, Object>> rowsAslistOfMaps =
+                      dataFetchingEnvironment.getArgument(tableMetadata.getIdentifier());
+                  if (rowsAslistOfMaps != null) {
+                    String tableName = tableMetadata.getTableName();
+                    Table table = tableMetadata.getTable();
+                    int count;
+                    List<Row> rows = TypeUtils.convertToRows(table.getMetadata(), rowsAslistOfMaps);
+                    switch (mutationType) {
+                      case UPDATE:
+                        count = table.update(rows);
+                        result.append("updated " + count + " records to " + tableName + "\n");
+                        break;
+                      case INSERT:
+                        count = table.insert(rows);
+                        result.append("inserted " + count + " records to " + tableName + "\n");
+                        break;
+                      case SAVE:
+                        count = table.save(rows);
+                        result.append("upserted " + count + " records to " + tableName + "\n");
+                        break;
+                      case DELETE:
+                        boolean strict =
+                            dataFetchingEnvironment.getArgumentOrDefault("strict", false);
+                        count = table.delete(rows, strict);
+                        result.append("delete " + count + " records from " + tableName + "\n");
+                        break;
+                    }
+                  }
+                }
+              });
       return new GraphqlApiMutationResult(SUCCESS, result.toString());
     };
   }
